Renamed source files with composite names
