+++ src/FeatureChecker.cpp
